0.3.3 (Mac version 3, PC version 3 Add Lock Computer to action on both versions Update version number for those early installers! You know who you are Fix for idletime in PC version Error check for missing items

+++ WinRemote.indigoPlugin/Contents/Server Plugin/plugin.py
@@ -690,13 +690,13 @@ class httpHandler(BaseHTTPRequestHandler):
                         MACaddress = dictparams['MAC']
                     idletime = 0
                     if 'Idle' in dictparams:
-                        idletime = int(dictparams['Idle'])
+                        idletime = float(dictparams['Idle'])
                     userName = 'unknown'
                     if 'userName' in dictparams:
                         userName = dictparams['userName']
                     upTime = 0
                     if 'upTime' in dictparams:
-                        upTime = int(dictparams['upTime'])
+                        upTime = float(dictparams['upTime'])
 
                     if dev.states['HostName']== dictparams['Hostname']:
                         #dev.updateStateOnServer('deviceIsOnline', value=True)
